Create an executor's scoped lock for the entire duration of the scope
TensorBufferScopedLock only provides a valid pointer to write to when
the object is in scope. Writing after the lock is destroyed can result
in undesired behavior.

During PrefillInternal(), the input_token's TensorBufferScopedLock
must stay in scope as long as the pointer is being written to.
The easiest way to ensure the TensorBufferScopedLock is still in scope
is to only create it when we are ready to write to it.

This also unifies more of the input_tensor vs input_embedding logic in
this function which will hopefully improve maintainability.

LiteRT-LM-PiperOrigin-RevId: 767209739